+++ projects/ids-enterprise-ng/src/lib/swipe-action/index.ts
@@ -1,0 +1,2 @@
+export * from './soho-swipe-action.component';
+export * from './soho-swipe-action.module';