+++ projects/ids-enterprise-ng/src/lib/swipe-action/soho-swipe-action.module.ts
@@ -1,0 +1,19 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import {
+  SohoSwipeActionComponent
+} from './soho-swipe-action.component';
+
+@NgModule({
+  imports: [
+    CommonModule,
+  ],
+  declarations: [
+    SohoSwipeActionComponent,
+  ],
+  exports: [
+    SohoSwipeActionComponent,
+  ],
+})
+export class SohoSwipeActionModule { }