+++ docs/CHANGELOG.md
@@ -17,6 +17,7 @@
 
 - `[ApplicationMenu]` Added Wrapper for Resizable Feature ([#1143](https://github.com/infor-design/enterprise-ng/issues/1143))
 - `[Message]` Added `showCloseBtn` setting to `SohoMessageOptions`. ([#1161](https://github.com/infor-design/enterprise-ng/issues/1161)) `EA`
+- `[Swipe Action]` Added a wrapper for Swipe Action Component. ([#1164](https://github.com/infor-design/enterprise-ng/issues/1164)) `EA`
 
 ## v10.9.0